Use data loaders to retrieve story, make hasRejectedAncestor respect archive
